rename conversationActionTypes to actionType
